Dynamic Routes workin and fetchin singular recipe by id
